group_donations/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.16
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-redis==5.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
drf-yasg==1.21.7
Faker==24.2.0
filetype==1.2.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
kombu==5.3.5
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
